+++ src/SubscriptionItem.php
@@ -2,7 +2,9 @@
 
 namespace Laravel\Cashier;
 
+use DateTimeInterface;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Collection;
 use Laravel\Cashier\Concerns\InteractsWithPaymentBehavior;
 use Laravel\Cashier\Concerns\Prorates;
 use Stripe\SubscriptionItem as StripeSubscriptionItem;
@@ -181,6 +183,37 @@ class SubscriptionItem extends Model
         $this->alwaysInvoice();
 
         return $this->swap($plan, $options);
+    }
+
+    /**
+     * Report usage for a metered product.
+     *
+     * @param  int  $quantity
+     * @param  \DateTimeInterface|int|null  $timestamp
+     * @return \Stripe\UsageRecord
+     */
+    public function reportUsage($quantity = 1, $timestamp = null)
+    {
+        $timestamp = $timestamp instanceof DateTimeInterface ? $timestamp->getTimestamp() : $timestamp;
+
+        return StripeSubscriptionItem::createUsageRecord($this->stripe_id, [
+            'quantity' => $quantity,
+            'action' => $timestamp ? 'set' : 'increment',
+            'timestamp' => $timestamp ?? time(),
+        ], $this->subscription->owner->stripeOptions());
+    }
+
+    /**
+     * Get the usage records for a metered product.
+     *
+     * @param  array  $options
+     * @return \Illuminate\Support\Collection
+     */
+    public function usageRecords($options = [])
+    {
+        return new Collection(StripeSubscriptionItem::allUsageRecordSummaries(
+            $this->stripe_id, $options, $this->subscription->owner->stripeOptions()
+        )->data);
     }
 
     /**